requirements.txt
python-can==4.2.2
textual==0.36.0
digi-xbee==1.4.1